runme.sh: change config for CEx7 TF-A build
The upstream TF-A uses different config for building
CEx7 firmware. Reflect that in the build script.

Additionally start using the official TF-A git
repo instead of the github mirror and change
the baseline to the one that includes CEx7 support.

Adjust the build command to produce the output
'flash-image.bin' binary - the latest TF-A
require explicit parameter for that purpose
('mrvl_flash').

Signed-off-by: Marcin Wojtas <mw@semihalf.com>