Add DPDK support.
Compile DPDK and MUSDK as part of the image.

Please note that once DPDK gain control over the network interfaces,
switching back to the Linux kernel is not supported.

If, for development purposes, switching back to the Linux kernel is
required, one can follow the following steps:

* Change the MTU for the wanted interface
ifconfig ethX mtu 1499
* Change the MTU back to the original value
ifconfig ethX mtu 1500

changing the MTU will reassociate the RXQ pools.
This hack is not stable, and will lead to error
messages from the kernel, and is not recommended in general.

Signed-off-by: Alvaro Karsz <alvaro.karsz@solid-run.com>