VPP patch - switch back control to kernel when VPP stops
New VPP command to close a DPDK interface, allowing to switch control
over the network interfaces back to the Linux kernel.

This command should be called before we exit VPP.

A user must not try to bring back the link again with VPP after calling
dpdk close, it will lead to vpp errors.

Signed-off-by: Alvaro Karsz <alvaro.karsz@solid-run.com>